+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.test.js
@@ -11,12 +11,20 @@ import userEvent from '@testing-library/user-event';
 
 import { pkg, carbon } from '../../settings';
 
-import { BreadcrumbItem, Button, Tab, Tabs } from 'carbon-components-react';
+import { BreadcrumbItem, Tab, Tabs } from 'carbon-components-react';
 import { Lightning16, Bee32 } from '@carbon/icons-react';
 
 import { PageHeader } from '.';
 import { ActionBarItem } from '../ActionBar';
-import { mockHTMLElement } from '../../global/js/utils/test-helper';
+import {
+  mockHTMLElement,
+  expectWarn,
+  expectMultipleWarn,
+  expectMultipleError,
+  deprecated,
+  deprecatedUsage,
+  required,
+} from '../../global/js/utils/test-helper';
 
 const { prefix } = pkg;
 
@@ -83,22 +91,6 @@ const pageActions = [
   },
 ];
 const pageActionsOverflowLabel = 'Page actions...';
-
-const pageActionsDepTest = pageActions.map(({ label, ...rest }, index) => (
-  <Button {...rest} key={index}>
-    {label}
-  </Button>
-));
-
-const pageActionsDepTest2 = (
-  <>
-    {pageActions.map(({ label, ...rest }, index) => (
-      <Button {...rest} key={index}>
-        {label}
-      </Button>
-    ))}
-  </>
-);
 
 const subtitle = 'Optional subtitle if necessary';
 const navigation = (
@@ -342,64 +334,15 @@ describe('PageHeader', () => {
     ).toHaveLength(1);
   });
 
-  test('copes with actionBarItems as nodes', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-
-    render(
-      <PageHeader
-        actionBarItems={actionBarItemsNodes}
-        actionBarOverflowAriaLabel={actionBarOverflowAriaLabel}
-      />
-    );
-
-    expect(warn).toBeCalledWith(
-      'The usage of the prop `actionBarItems` of `PageHeader` has been changed and support for the old usage will soon be removed. Expects an array of objects with the following properties: iconDescription, renderIcon and onClick.'
-    );
-
-    warn.mockRestore(); // Remove mock
-  });
-
-  test('with deprecated page actions', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-
-    render(
-      <PageHeader
-        pageActions={pageActionsDepTest}
-        pageActionsOverflowLabel={pageActionsOverflowLabel}
-      />
-    );
-
-    screen.getByText('Primary button', {
-      selector: `.${blockClass}__page-actions .${prefix}--button-set-with-overflow__button-container:not(.${prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
-    });
-
-    expect(warn).toBeCalledWith(
-      'The usage of the prop `pageActions` of `PageHeader` has been changed and support for the old usage will soon be removed. Expects an array of objects with the following properties: label and onClick.'
-    );
-
-    warn.mockRestore(); // Remove mock
-  });
-
-  test('with deprecated page actions in a fragment', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-
-    render(
-      <PageHeader
-        pageActions={pageActionsDepTest2}
-        pageActionsOverflowLabel={pageActionsOverflowLabel}
-      />
-    );
-
-    screen.getByText('Primary button', {
-      selector: `.${blockClass}__page-actions .${prefix}--button-set-with-overflow__button-container:not(.${prefix}--button-set-with-overflow__button-container--hidden) .${carbon.prefix}--btn`,
-    });
-
-    expect(warn).toBeCalledWith(
-      'The usage of the prop `pageActions` of `PageHeader` has been changed and support for the old usage will soon be removed. Expects an array of objects with the following properties: label and onClick.'
-    );
-
-    warn.mockRestore(); // Remove mock
-  });
+  test('copes with actionBarItems as nodes', () =>
+    expectWarn(deprecatedUsage('actionBarItems', 'PageHeader'), () => {
+      render(
+        <PageHeader
+          actionBarItems={actionBarItemsNodes}
+          actionBarOverflowAriaLabel={actionBarOverflowAriaLabel}
+        />
+      );
+    }));
 
   const dataTestId = 'data-testid';
 
@@ -628,28 +571,25 @@ describe('PageHeader', () => {
     screen.getByRole('region');
   });
 
-  test('ActionBar without overflow aria label', () => {
-    const error = jest.spyOn(console, 'error').mockImplementation(() => {});
-
-    const { title } = testProps;
-    render(
-      <PageHeader
-        {...{
-          title,
-          actionBarItems,
-        }}
-        aria-label="Page header" // gives section role 'region'
-      />
-    );
-
-    expect(error).toBeCalledWith(
-      expect.stringMatching(
-        /^Warning: Failed prop type: The prop `overflowAriaLabel` is marked as required in `ActionBar`/
-      )
-    );
-
-    jest.spyOn(console, 'error').mockRestore();
-  });
+  test('ActionBar without overflow aria label', () =>
+    expectMultipleError(
+      [
+        required('actionBarOverflowAriaLabel', 'PageHeader'),
+        required('overflowAriaLabel', 'ActionBar'),
+      ],
+      () => {
+        const { title } = testProps;
+        render(
+          <PageHeader
+            {...{
+              title,
+              actionBarItems,
+            }}
+            aria-label="Page header" // gives section role 'region'
+          />
+        );
+      }
+    ));
 
   test('Title shows as loading', () => {
     render(